pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fedctl"
version = "0.0.1"
description = "Federated control CLI"
requires-python = ">=3.9"
dependencies = [
  "typer>=0.12.0",
  "tomlkit>=0.12.0",
  "rich>=13.0.0",
  "httpx>=0.27.0",
  "jinja2>=3.1.0",
  "PyYAML>=6.0.0",
]

[project.scripts]
fedctl = "fedctl.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
